Fix unhandled promise rejection on emergency unlinking
When an error occurs, we clean up the temporary file on the file system through
an unlinking. However, prior to this patch, the unlinking's potential promise
rejection was not handled. This patch fixes that.

This change, along with the previous commit migrating to a native promise when
available, should fix #3.